requirements.txt
alembic==1.4.3
apache-airflow==1.10.9
apispec==1.3.3
argcomplete==1.12.2
attrs==20.3.0
Babel==2.9.0
bcrypt==3.2.0
cached-property==1.5.2
cachetools==4.2.0
cattrs==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.37
defusedxml==0.6.0
dill==0.3.3
dnspython==2.1.0
docutils==0.16
duckdb==0.2.8
durable-rules==2.0.28
email-validator==1.1.2
grpcio==1.40.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
importlib-metadata==2.1.1
importlib-resources==1.5.0
iso8601==0.1.13
itsdangerous==1.1.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
natsort==7.1.0
numpy==1.19.5
oauthlib==3.1.0
packaging==20.8
pandas==1.1.5
pendulum==1.4.4
pep562==1.0
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2020.5
pytzdata==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
setproctitle==1.2.1
six==1.15.0
SQLAlchemy==1.3.15
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
starkbank-ecdsa==1.0.0
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.3
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==0.16.1
zipp==3.4.0
zope.deprecation==4.4.0
